Extracted DescribedUpdateSite.Descriptor to DescribedUpdateSiteDescriptor.
